ai-orchestration/requirements.txt
openai==2.29.0
openai-agents==0.13.0
python-dotenv==1.2.2
